feat: move env to config
