+++ src/config/index.ts
@@ -1,6 +1,3 @@
-import { env } from "~/env";
-
-export const isLocal = env.VITE_APP_ENV === "local";
-
 export * from "./common";
+export * from "./env";
 export * from "./storybook";
