+++ src/services/api/axios.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-import { env } from "~/env";
+import { env } from "~/config";
 import {
   authHeaderInterceptor,
   errorResponseInterceptor,
